requirements.txt
transformers>=4.35.0
torch>=2.0.0
Pillow>=10.0.0
accelerate>=0.20.0
langgraph>=0.0.20
langfuse==2.59.7
langchain